fix: address Gemini code review feedback
- Extract model inference logic into inferModelNameFromAgent() in model-capabilities.ts
- Remove dead onClick handler on disabled mic button (tooltip suffices)
- Reuse hasValidTranscript instead of redundant isValidTranscript call
- Add changeset for version bump
