+++ .changeset/model-voice-input.md
@@ -1,0 +1,5 @@
+---
+"@iqai/adk": patch
+---
+
+feat: add model-aware voice input with audio support detection
